21 filter sort, need to fix search entry, update inStock and sold

+++ pages/api/categories/[id].js
@@ -1,6 +1,7 @@
 /* eslint-disable import/no-anonymous-default-export */
 import connectDB from '../../../utils/connectDB'
 import Categories from '../../../models/categoriesModel'
+import Products from '../../../models/productModel'
 import auth from '../../../middleware/auth'
 
 
@@ -44,7 +45,10 @@ const deleteCategory = async (req, res) => {
         if(result.role !== 'admin') return res.status(400).json({err: 'Authentication is not valid'})
 
         const {id} = req.query
-        const {name} = req.body
+
+        const products = await Products.findOne({category: id})
+        if(products) return res.status(400).json({err: "Please delete all products with a relathionship"})
+
 
         await Categories.findByIdAndDelete(id)
